+++ src/main/java/vendingmachine/service/VendingMachineService.java
@@ -1,7 +1,6 @@
 package vendingmachine.service;
 
 import vendingmachine.constants.ErrorMessage;
-import vendingmachine.model.Coin;
 import vendingmachine.model.VendingMachine;
 import vendingmachine.view.UserView;
 import vendingmachine.view.VendingMachineView;
@@ -45,20 +44,6 @@ public class VendingMachineService {
 			return false;
 		}
 		return true;
-	}
-
-	public void giveChanges() {
-		vendingMachineView.printChangesComment(vendingMachine.getRemainMoney());
-
-		for (Coin coin : Coin.values()) {
-			int changes = vendingMachine.giveChanges(coin);
-
-			if (changes != 0) {
-				vendingMachineView.printCoin(changes, coin);
-			}
-
-		}
-
 	}
 
 }
